refactor: centralize offline update
